requirement.txt
dgl-cu117==0.9.1.post1
Pillow==9.5.0
scikit-learn==1.0.2
scipy==1.7.3
torch==1.13.1
torchvision==0.14.1
tqdm==4.66.2
